+++ urbanairship-message-center/src/test/java/com/urbanairship/messagecenter/MessageDaoTest.java
@@ -19,8 +19,10 @@ import static org.junit.Assert.assertEquals;
 
 @RunWith(AndroidJUnit4.class)
 public class MessageDaoTest {
-    private String messageId = "1_message_id";
-    private String messageId2 = "2_message_id";
+    private final String messageId = "1_message_id";
+    private final String messageId2 = "2_message_id";
+    private final String messageTitle1 = "Message 1 Title";
+    private final String messageTitle2 = "Message 2 Title";
     private JsonValue messageJson;
     private JsonValue messageJson2;
 
@@ -40,7 +42,7 @@ public class MessageDaoTest {
                 "                 \"message_id\": \"" + messageId + "\"" +
                 "               }," +
                 "\"unread\": true, \"message_sent\": \"2010-09-05 12:13 -0000\"," +
-                "\"title\": \"Message title\", \"extra\": { \"some_key\": \"some_value\"}," +
+                "\"title\": \"" + messageTitle1 + "\", \"extra\": { \"some_key\": \"some_value\"}," +
                 "\"content_type\": \"text/html\", \"content_size\": \"128\"}");
         messageJson2 = JsonValue.parseString("{\"message_id\": \"" + messageId2 + "\"," +
                 "\"message_url\": \"https://go.urbanairship.com/api/user/userId/messages/message/some_mesg_id/\"," +
@@ -50,13 +52,16 @@ public class MessageDaoTest {
                 "                 \"message_id\": \"" + messageId2 + "\"" +
                 "               }," +
                 "\"unread\": true, \"message_sent\": \"2010-09-05 12:13 -0000\"," +
-                "\"title\": \"Message title\", \"extra\": { \"some_key\": \"some_value\"}," +
+                "\"title\": \"" + messageTitle2 + "\", \"extra\": { \"some_key\": \"some_value\"}," +
                 "\"content_type\": \"text/html\", \"content_size\": \"128\"}");
-        db = MessageDatabase.createInMemoryDatabase(ApplicationProvider.getApplicationContext());
-        messageDao = db.getDao();
 
         ENTITY = MessageEntity.createMessageFromPayload(messageId, messageJson);
         ENTITY2 = MessageEntity.createMessageFromPayload(messageId2, messageJson2);
+
+        db = MessageDatabase.createInMemoryDatabase(ApplicationProvider.getApplicationContext());
+        messageDao = db.getDao();
+
+        assertEquals(0, messageDao.getMessages().size());
     }
 
     @After
@@ -66,14 +71,26 @@ public class MessageDaoTest {
 
     @Test
     public void testInsert() {
-        assertEquals(0, messageDao.getMessages().size());
         messageDao.insert(ENTITY);
         assertEquals(1, messageDao.getMessages().size());
     }
 
+    @Test()
+    public void testInsertsReplaceOnConflict() {
+        // Insert duplicate messages.
+        insertMessage("foo", messageJson);
+        insertMessage("foo", messageJson);
+        assertEquals(1, messageDao.getMessages().size());
+        assertEquals(messageTitle1, messageDao.getMessages().get(0).title);
+
+        // Insert another with same ID, but different data.
+        insertMessage("foo", messageJson2);
+        assertEquals(1, messageDao.getMessages().size());
+        assertEquals(messageTitle2, messageDao.getMessages().get(0).title);
+    }
+
     @Test
     public void testGetMessages() {
-        assertEquals(0, messageDao.getMessages().size());
         messageDao.insert(ENTITY);
         messageDao.insert(ENTITY2);
         assertEquals(2, messageDao.getMessages().size());
@@ -113,8 +130,6 @@ public class MessageDaoTest {
 
     @Test
     public void testDeleteTooManyMessages() {
-        assertEquals(0, messageDao.getMessages().size());
-
         List<String> messageIds = insertMessages(2000);
         assertEquals(2000, messageDao.getMessages().size());
 
@@ -122,48 +137,21 @@ public class MessageDaoTest {
         assertEquals(0, messageDao.getMessages().size());
     }
 
-    @Test
-    public void testDeleteTooManyDuplicateMessages() {
-        assertEquals(0, messageDao.getMessages().size());
-
-        List<String> messageIds = insertDuplicateMessages(2000);
-        assertEquals(2000, messageDao.getMessages().size());
-
-        messageDao.deleteDuplicates();
-        assertEquals(1, messageDao.getMessages().size());
-    }
-
     @SuppressWarnings("SameParameterValue")
     private List<String> insertMessages(int count) {
-        List<MessageEntity> messages = new ArrayList<>();
         List<String> messageIds = new ArrayList<>();
 
         for (int i = 0; i < count; i++) {
-            MessageEntity message = MessageEntity.createMessageFromPayload(UUID.randomUUID().toString(), messageJson);
-            assert message != null;
-            messages.add(message);
-            messageIds.add(message.getMessageId());
+            messageIds.add(insertMessage(UUID.randomUUID().toString(), messageJson));
         }
-
-        messageDao.insertMessages(messages);
 
         return messageIds;
     }
 
-    @SuppressWarnings("SameParameterValue")
-    private List<String> insertDuplicateMessages(int count) {
-        List<MessageEntity> messages = new ArrayList<>();
-        List<String> messageIds = new ArrayList<>();
-
-        for (int i = 0; i < count; i++) {
-            MessageEntity message = MessageEntity.createMessageFromPayload("theExactSameId", messageJson);
-            assert message != null;
-            messages.add(message);
-            messageIds.add(message.getMessageId());
-        }
-
-        messageDao.insertMessages(messages);
-
-        return messageIds;
+    private String insertMessage(String id, JsonValue json) {
+        MessageEntity message = MessageEntity.createMessageFromPayload(id, json);
+        assert message != null;
+        messageDao.insert(message);
+        return message.getMessageId();
     }
 }